# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Enterprise API",
        "description": "API for managing enterprise subscriptions and users",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/plans": {
            "get": {
                "tags": [
                    "Plans"
                ],
                "summary": "Get list of plans",
                "operationId": "8b6d4b7c567837ca8acb89185ccdc8e7",
                "responses": {
                    "200": {
                        "description": "List of plans retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Basic Plan"
                                                    },
                                                    "monthly_price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 9.99
                                                    },
                                                    "user_limit": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "features": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Feature 1"
                                                                },
                                                                "description": {
                                                                    "type": "string",
                                                                    "example": "Description of feature 1"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Plans retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Plans"
                ],
                "summary": "Create a new plan",
                "operationId": "532452d7daf47841b565ae3468545c16",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "monthly_price",
                                    "user_limit",
                                    "features"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Basic Plan"
                                    },
                                    "monthly_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 9.99
                                    },
                                    "user_limit": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "features": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Feature 1"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Description of feature 1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plan created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Basic Plan"
                                                },
                                                "monthly_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 9.99
                                                },
                                                "user_limit": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "features": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Feature 1"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Description of feature 1"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Plan created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name field is required"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearer_token": []
        }
    ]
}